ExploreWithMe.main-svc: Events - начало

+++ main-svc/src/main/java/ru/yandex/practicum/event/dto/EventFullDto.java
@@ -35,5 +35,5 @@ public class EventFullDto {
     private Boolean requestModeration;
     private String state;
     private String title;
-    private Integer views;
+    private Long views;
 }
